+++ benchmarks/profiler/profile_sla.py
@@ -22,11 +22,8 @@ import os
 import numpy as np
 import yaml
 
-from benchmarks.profiler.utils.config import (
-    CONFIG_MODIFIERS,
-    WORKER_COMPONENT_NAMES,
-    generate_dgd_config_with_planner,
-)
+from benchmarks.profiler.utils.config import generate_dgd_config_with_planner
+from benchmarks.profiler.utils.config_modifiers import CONFIG_MODIFIERS
 from benchmarks.profiler.utils.estimate_perf import AIConfiguratorPerfEstimator
 from benchmarks.profiler.utils.genai_perf import benchmark_decode, benchmark_prefill
 from benchmarks.profiler.utils.planner_utils import add_planner_arguments_to_parser
@@ -53,6 +50,7 @@ from deploy.utils.dynamo_deployment import (
     DynamoDeploymentClient,
     cleanup_remaining_deployments,
 )
+from dynamo.planner.defaults import WORKER_COMPONENT_NAMES
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
